Fix bug when too small image gets cropped before Profile screen started, so broadcast is lost because receiver was not registered.
Introduced global receiver app-wide, that should receive image cropping
results and then pass it to local receiver on Profile screen, when started
and local receiver is registered. Global receiver is then unregistered.

But on logout, it should re-register global receiver again, because app
does not restart and global register is unsubscried since start of Profile
screen.