Fix small images not show on Profile
Previous assumption was not quite correct - broadcast intent is not STICKY
in CropIwa library, so it can not be received by receivers that have
registered after that intent had been sent. So only app-wide and CropIwa
internal receivers actually receive that intent, and for small images
cropping is too fast, so local receiver on UserProfile screen is
registered after the intent has been sent, so it does not receive it at
all (because intent is not STICKY, as mentioned).

So we just need to subscribe on receives on app-wide rerceiver and react
on them. App-wide receiver doesn't require to be unregistered.